Add monotonous autogluon training.
The only supported models are: XGBoost, CatBoost, and a monotonic
dense neural network implemented in PyTorch (I have not yet implemented
a FastAI monotonic neural network).

There is a parameter for monotonic constraints for the LightGBM model,
but in practice it does not produce monotonic functions. I removed
the model from autogluon's pool of models.