Bug branch for monotonous model loading error.
This branch is created to try and fix the error
"ModuleNotFoundError: No module named 'monotone_network_autogluon'"
that occurs when I try to evaluate the trained monotone neural network.

The cause of the bug seems to be that the file is not in the path
when the model is reloaded after being saved.