Set monotone constraints as param for the network.
The monotone constraints on the features are no longer initialised in
the __init__ of the MonoDenseTorch class. Instead, the constraints are
passed as parameters for the forward method. To adhere with the
torch.nn.Module API, the constraints are bundled with the features when
they are passed as a parameter, and unpacked inside the function.

Moreover, a useless casting of a torch.Tensor into a torch.Tensor is
removed. The typing is fixed accordingly.