Fix monotone constraint retention.
Updates the monotonous internals of autogluon by setting the monotone
constraints as an attribute of the MonotoneAutoTrainer. This allows the
retention of the constraints from the fitting of the Trainer to the
fitting of a model (e.g., an XGBoost model).

Earlier, I tried to pass the monotone constraints as arguments through
the stack of functions (train_multi_levels, stack_new_level, etc.) but
the constraints are lost somewhere during the transmission.