Add MonotonousMultiLabelPredictor.
Adds a monotonous multi label predictor that relies on
MonotonePredictors to compute individual constraints. Note that for the
cost predictor, we should probably not enforce monotone constraints. If
we increase the number of cores, we increase the cost, but if the
training is faster, we reduce the cost. Thus, the monotone constraints
may not apply in this setting.