Fix issue with input dimension of transformer.
Fixes an issue with the pre-processing of the query graph, where the
dimension of the inputs and the Linear layers were different.

The type dimension had not been added to the input dimension of the
Linear layer, so the Linear layer had a smaller input dimension than
what was required.